#review-95202
ECPAPPSERVERWEBLOGIC-32
o Updated the file references for createDataSource and deleteDataSource in the plugin's manifest.pl file. There was a case mis-match in the file names due to which the plugin jars built on a *nix system did not include the files required for the createDataSource and deleteDataSource procedures.
o Also updated the buildProject-mvn.pl script to detect such missing files and report it as a warning. The build below shows the warnings for all those plugins which have incorrect file references. 
https://chronic3/commander/link/jobDetails/jobs/1f82c1f3-ae35-11e4-8705-02bfc0a80055?objectId=job-1f82c1f3-ae35-11e4-8705-02bfc0a80055
We will file a separate JIRA to fix these references and also convert the warning to an error so that such errors are caught early going forward.

[git-p4: depot-paths = "//depot/nimbus/integrations/EC-WebLogic/main/": change = 95201]